+++ src/Magento/FunctionalTestingFramework/Test/Objects/ActionGroupObject.php
@@ -25,7 +25,7 @@ class ActionGroupObject
         "executeJS",
         "magentoCLI",
         "generateDate",
-        "formatMoney",
+        "formatCurrency",
         "deleteData",
         "getData",
         "updateData",